+++ CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Calendar Versioning](https://calver.org).
 ### Experimental
 
 ### Changed
+- [CSE] Allow virtual resources to be referenced in &lt;actions> and &lt;dependencies>.
 - [TUI] Diagrams now support containers that contain boolean-kind values (true/false, on/off, yes/no, high/low).
 
 ### Fixed